Moved Serial type defintions to a new model file; updated related code as needed
